simulator: network delays are now 1-way, but proximity() is the RTT, call networkDelay() for the uni-directional distance   reduced some rounding problems related to proximity()/networkDelay().  Which was causing delay streatch of < 1 in a euclidean network.  ... networkDelay() is now a double.
Fixed some bugs with simulating churn in the simulator:
  made PastryNode.destroy() work better, including making MessageDispatch a Destructable, and the ability to add destructables to the PN.

if all bootstraps go faulty while joining, node will:
  a) send a JoinFailedException to the observers, 
  b) calling PN.joinFailed() will return true.
  c) PN.notifyAll() will be called, so code waiting on the PN can wake up and see how to handle it (can call joinFailed() to see if that is why they were notified).

git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@3413 c21dacaf-8602-0410-aacd-8bc41df2cbe7

+++ pastry/src/rice/pastry/direct/BasicNetworkSimulator.java
@@ -323,7 +323,7 @@ public abstract class BasicNetworkSimulator implements NetworkSimulator {
    * @param b the second NodeId
    * @return the proximity between the two input NodeIds
    */
-  public int networkDelay(DirectNodeHandle a, DirectNodeHandle b) {
+  public double networkDelay(DirectNodeHandle a, DirectNodeHandle b) {
     NodeRecord nra = a.getRemote().record;
     NodeRecord nrb = b.getRemote().record;
